+++ docs/reference/coordinates.rst
@@ -1,0 +1,12 @@
+.. _coordinates:
+
+Coordinates ('stixpy.coordinates')
+**********************************
+
+The ``coordinates`` submodule contains STIX specific coordinate frames and transforms.
+
+.. automodapi:: stixpy.coordinates
+
+.. automodapi:: stixpy.coordinates.frames
+
+.. automodapi:: stixpy.coordinates.transforms
