--- docs/reference/frames.rst
@@ -1,8 +1,0 @@
-.. _frames:
-
-Frames ('stixpy.frames')
-************************
-
-The ``frames`` submodule contains STIX specific coordinate frames and transforms.
-
-.. automodapi:: stixpy.frames
