+++ docs/reference/index.rst
@@ -7,12 +7,15 @@ Reference
 .. toctree::
    :maxdepth: 2
 
+   coordinates
    data
    calibration
    timeseries
    science
+   map
    product
+   science
+   timeseries
    visualisation
-   frames
 
    ../whatsnew/index
