+++ docs/reference/map.rst
@@ -1,0 +1,10 @@
+.. _map:
+
+Map ('stixpy.map')
+******************
+
+The ``map`` submodule contains stix specific map source for `~sunpy.map.Map`
+
+.. automodapi:: stixpy.map
+
+.. automodapi:: stixpy.map.stix
