+++ setup.cfg
@@ -68,7 +68,7 @@ filterwarnings =
 
 [isort]
 balanced_wrapping = true
-skip =
+extend_skip =
     docs/conf.py,
     stixpy/__init__.py,
 default_section = THIRDPARTY
